pyproject.toml
[tool.poetry]
name = "interview"
version = "0.1.0"
description = ""
authors = ["aowerdmax <qq1329210652@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aliyun-python-sdk-core = "^2.15.1"
python-dotenv = "^1.0.1"
websockets = "^12.0"
aiohttp = "^3.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true
